update types to be able to hold relevant info and noticed a data read function that needs new global info

+++ contracts/StakeableToken.sol
@@ -279,7 +279,7 @@ contract StakeableToken is UTXORedeemableToken {
      * @param sharesPerHeart number of shares for each heart
      */
 
-    function calcStakeShares(uint256 newStakedHearts, uint256 sharesPerHeart)
+    function calcStakeShares(uint256 newStakedHearts, uint80 sharesPerHeart)
         private
         pure
         returns (uint256)
